requirements.txt
Appium-Python-Client==0.49
certifi==2019.11.28
chardet==3.0.4
idna==2.8
PyYAML==5.3
requests==2.22.0
selenium==3.141.0
urllib3==1.25.7
